googledino/pip-requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
Pillow
# https://python-mss.readthedocs.io/usage.html -- different code for different platforms
mss==7.0.1

# https://pyautogui.readthedocs.io/en/latest/install.html
pyautogui
